chore: add docs command to be used by mage
